+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   * `rustfmt --check` now run by default
 * Fix `#` format when not used as a last argument.
 * Implement `Value` for `std::borrow::Cow`
+* Fix duplicate `AsRef<str>` implementations, fixing support for `dynamic-keys`
 
 ## 2.7.0 - 2020-11-29